Use CMake for platform specific cpu compile
