+++ mlx/compile.cpp
@@ -7,6 +7,7 @@
 
 #include "mlx/allocator.h"
 #include "mlx/compile.h"
+#include "mlx/compile_impl.h"
 #include "mlx/primitives.h"
 #include "mlx/transforms.h"
 #include "mlx/transforms_impl.h"
@@ -766,7 +767,7 @@ std::function<std::vector<array>(const std::vector<array>&)> compile(
     bool shapeless /* = false */,
     std::vector<uint64_t> constants /* = {} */) {
   if (compile_mode() == CompileMode::disabled ||
-      (default_device() == Device::cpu && !CPU_COMPILE)) {
+      !(compile_available_for_device(default_device()))) {
     return fun;
   }
   return [fun, fun_id, shapeless, constants = std::move(constants)](
